pyproject.toml
[tool.poetry]
name = "auth-system"
version = "0.1.0"
description = "App to authenticate persons"
authors = ["Albert Weihermann"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.3"
libsass = "^0.21.0"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.3"
bcrypt = "^3.2.0"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
prospector = "^1.7.5"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
